% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_sort.R
\name{create_hoare_partition}
\alias{create_hoare_partition}
\title{create_hoare_partition}
\usage{
create_hoare_partition(V)
}
\arguments{
\item{V}{a vector of numeric data}
}
\value{
a list containing the index of the randomly selected pivot after partitioning and the vector after partitioning
}
\description{
Apply Hoare partitioning on a vector
}
\examples{
create_hoare_partition(V = 1:10)
}
