% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_sort.R
\name{quick_sort}
\alias{quick_sort}
\title{quick_sort}
\usage{
quick_sort(V)
}
\arguments{
\item{V}{the vector to sort}
}
\value{
the sorted vector
}
\description{
Quick sort algorithm using Hoare partitioning
}
\examples{
quick_sort(V = sample(10))
}
