% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_sort.R
\name{quick_sort_opti}
\alias{quick_sort_opti}
\title{quick_sort_opti}
\usage{
quick_sort_opti(V)
}
\arguments{
\item{V}{the vector to sort}
}
\value{
the sorted vector
}
\description{
Quick sort algorithm using three way partitioning
}
\examples{
quick_sort_opti(V = sample(10))
}
