% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radix_sort.R
\name{tri_digit}
\alias{tri_digit}
\title{tri_digit}
\usage{
tri_digit(V, rank)
}
\arguments{
\item{V}{le vecteur à trier selon l'indicateur de position rank}

\item{rank}{indicateur de position (1 pour les unités, 2 pour les dizaines, etc.)}
}
\value{
Le vecteur trié selon l'indicateur de position rank
}
\description{
Trie le vecteur V selon un indicateur de position nommé rank (1 pour trier selon les unités, 2 pour trier selon les dizaines et les unités etc.)
}
