% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radix_sort.R
\name{tri_digit_opti}
\alias{tri_digit_opti}
\title{tri_digit}
\usage{
tri_digit_opti(V, rank)
}
\arguments{
\item{V}{le vecteur à trier selon l'indicateur de position rank}

\item{rank}{indicateur de position (1 pour les unités, 2 pour les dizaines, etc.)}
}
\value{
Le vecteur trié selon l'indicateur de position rank
}
\description{
Version optimisée de la fonction tri_digit
}
